requirements.txt
opencv-python==4.8.1.78
mediapipe==0.10.8
ultralytics==8.0.205
numpy==1.24.3
pynput==1.7.6
Pillow==10.0.1
matplotlib==3.7.2
PyYAML==6.0.1
configparser==6.0.0
customtkinter==5.2.0
pytest==7.4.2
black==23.9.1
flake8==6.1.0